refactor smiles_from_pdb and add tests
